Move deprecation to method level. Remove redundant package level.
