+++ src/main/java/no/ndla/taxonomy/rest/v1/dtos/NodeResourceDTO.java
@@ -1,4 +1,11 @@
-package no.ndla.taxonomy.rest.v1.dtos.nodes;
+/*
+ * Part of NDLA taxonomy-api
+ * Copyright (C) 2023 NDLA
+ *
+ * See LICENSE
+ */
+
+package no.ndla.taxonomy.rest.v1.dtos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
