+++ src/main/java/no/ndla/taxonomy/rest/v1/dtos/searchapi/LanguageFieldDTO.java
@@ -1,4 +1,11 @@
-package no.ndla.taxonomy.rest.v1.dtos.nodes.searchapi;
+/*
+ * Part of NDLA taxonomy-api
+ * Copyright (C) 2023 NDLA
+ *
+ * See LICENSE
+ */
+
+package no.ndla.taxonomy.rest.v1.dtos.searchapi;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import no.ndla.taxonomy.domain.LanguageField;
