+++ src/main/java/no/ndla/taxonomy/rest/v1/dtos/searchapi/SearchableTaxonomyResourceType.java
@@ -1,4 +1,11 @@
-package no.ndla.taxonomy.rest.v1.dtos.nodes.searchapi;
+/*
+ * Part of NDLA taxonomy-api
+ * Copyright (C) 2023 NDLA
+ *
+ * See LICENSE
+ */
+
+package no.ndla.taxonomy.rest.v1.dtos.searchapi;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import no.ndla.taxonomy.domain.ResourceType;
